+++ changelog.md
@@ -2,6 +2,7 @@
 
 ####/Manhunt2.DebugMenu/dllmain.cpp
 * Issue: Core Feature 1 - 3 State Execution Firearm
+* Issue: Core Feature 2 - Normal headshots
 
 
 ####/global/resource1.glg